% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seurat_metadata.R
\name{add_seurat_metadata}
\alias{add_seurat_metadata}
\title{Define 10x clonotypes}
\usage{
add_seurat_metadata(
  seu = NULL,
  clonal.list = ls(pattern = "^Clonal", envir = .GlobalEnv),
  cdr3_map = "cdr3_col2",
  cell = "T",
  purity = 0.8,
  stick_only = "all",
  sticky = FALSE,
  overwrite = T
)
}
\arguments{
\item{seu}{Seurat. A seurat object to embed with clonality data.}

\item{clonal.list}{Character. The clonalities to be inserted into the seurat metadata. Default:  \code{ls(pattern = "^Clonal")}.}

\item{cdr3_map}{Character. The columun name to be used to map single chains to paired chains. CDR3 for nt ; cdr3_col2 for aa. Default: "cdr3_col2".}

\item{cell}{Character. Possible values: \code{B} Bcells, \code{T} Tcells, \code{Tgd} Tcells GamaDelta.}

\item{purity}{Numeric. Minimum fraction of the dominant mapped paired to be used in a sticky assignment.}

\item{stick_only}{Character vector. Possible values: \code{TRA}, \code{TRB}, \code{TRD}, \code{TRG}, \code{IGH}, \code{IGK}, \code{IGL}, \code{all}. Default: \code{all}}

\item{sticky}{Logical. If the script should merge the clonality of single chain cells with paired cells.}

\item{overwrite}{Logical. Whether to overwrite clonality columns from seurat.}
}
\description{
Takes a list of paired chains and defines clonotypes.
}
