% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigntenx.R
\name{assigntenx}
\alias{assigntenx}
\title{Define 10x clonotypes}
\usage{
assigntenx(
  list.pairs = list.pairs,
  method = method,
  clonality_input = clonality_input,
  cell = cell
)
}
\arguments{
\item{list.pairs}{List. List of cells generated by tenx.}

\item{clonality_input}{Named vector. Changes the parameters of clonality function.}

\item{cell}{Character. Possible values: \code{B} Bcells, \code{T} Tcells, \code{Tgd} Tcells GamaDelta.}
}
\description{
Takes a list of paired chains and defines clonotypes.
}
