% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_version_parser.R
\name{cellranger_version_parser}
\alias{cellranger_version_parser}
\title{Parse and Validate Cell Ranger Data}
\usage{
cellranger_version_parser(data, cellranger_version = 7)
}
\arguments{
\item{data}{Either a data frame, the name of a data object, or a file path to a \code{.csv} file.}

\item{cellranger_version}{Integer. Must be either 3 or 7. Used to validate the expected number of columns. Default is 7.}
}
\value{
A validated data frame with structure compatible with the specified Cell Ranger version.
}
\description{
Accepts a data frame, a variable name as a character string, or a file path to a \code{.csv} file.
Returns a parsed data frame and validates the number of columns according to Cell Ranger version.
}
\examples{
parsed <- cellranger_version_parser(Cellranger3_TCR, cellranger_version = 3)
parsed <- cellranger_version_parser("Cellranger3_TCR", cellranger_version = 3)

}
