% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{clonality}
\alias{clonality}
\title{Define clonality}
\usage{
clonality(
  data = "example.xlsx",
  output = "output",
  id_col = "Sequence_ID",
  vgene_col = "V_GENE_and_allele",
  jgene_col = "J_GENE_and_allele",
  cdr3_col = "JUNCTION",
  cell = "T",
  rm.na = TRUE,
  output_original = FALSE,
  mm = 0,
  suffix = NULL,
  search_gname = T
)
}
\arguments{
\item{data}{Data frame object or the full path to a xlsx with rep data.}

\item{output}{The data output name. Default: output.}

\item{id_col}{The column with an unique ID. Default: 'Sequence_ID'.}

\item{vgene_col}{The column with V gene names. Default: 'V_GENE_and_allele'.}

\item{jgene_col}{The column with J gene names. Default: 'J_GENE_and_allele'.}

\item{cdr3_col}{The column with CDR3 sequences - nt/aa. Default: 'JUNCTION'.}

\item{cell}{Choose between T for T cell or B for B cell. Default: 'T'.}

\item{rm.na}{Remove NA junctions. If \code{FALSE}, set to unique. Default: 'TRUE'.}

\item{output_original}{If \code{TRUE}, output = input with extra clonality column. Default: 'FALSE'.}

\item{mm}{Percent of mismatches allowed in subgroups. Default: 0}
}
\description{
Returns table with clonality definitions.
}
\examples{
clonality(data = tra)
clonality(data = trb)
clonality(data = 'example.xlsx')
}
