% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.rem.r
\name{clonEM}
\alias{clonEM}
\title{Auxiliary EM algorithm Function.}
\usage{
clonEM(mutmat, init.para, xigrid, conv.crit, niter)
}
\arguments{
\item{mutmat}{Matrix containing the data, with all mutations in rows and the tumor pairs in columns. 
The data are coded as 0=mutation not observed, 1=shared mutation (observed in both tumors), 2=private mutation 
(observed in one tumor only). The first column contains the probabilities of occurence for each mutation.}

\item{init.para}{Initial values of the parameters for the optimization. The order of the parameters is 
c(mu, sigma, pi), where mu and sigma are the mean and variance of the lognormal distribution 
of the random-effect xi, and pi is the proportion of clonal cases.}

\item{xigrid}{Grid of the values of xi used to compute the integration; it corresponds to the domain of 
definition of xi. }

\item{conv.crit}{Criteria used to defined convergence (on all three parameters).}

\item{niter}{Maximum number of iterations used for the EM algorithm.}
}
\value{
Return the value of the parameters maximizing the likelihood, the number of iterations used, and the convergence status.
}
\description{
This function optimizes the likelihood of the model using an EM algorithm.
}


\author{
Audrey Mauguen \email{mauguena@mskcc.org} and Venkatraman E. Seshan.
}


