% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequences_imgt.R
\name{filter_sequences_imgt}
\alias{filter_sequences_imgt}
\title{Filter Bad Sequences}
\usage{
filter_sequences_imgt(df, cdr3_col)
}
\arguments{
\item{df}{A data frame containing repertoire data.}

\item{cdr3_col}{\code{Character}. The column name containing CDR3 sequences.}
}
\value{
A filtered data frame with the bad sequences removed and a message reporting the number of rows removed.
}
\description{
This function filters out bad sequences from the input data frame based on the following criteria:
\itemize{
\item Rows with missing CDR3 sequences (\code{NA}, \code{""}).
\item Non-productive sequences (if a column containing "Functionality" is found, only rows with values like \code{TRUE} or \code{productive} are kept).
\item CDR3 sequences containing invalid characters (e.g., \code{*}, \verb{#}) if they are amino acid sequences (non-nucleotide).
}
}
\examples{
filtered_df <- filter_sequences_imgt(df, cdr3_col = "JUNCTION")
}
