% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.lik.r
\name{grid.lik}
\alias{grid.lik}
\title{Auxiliary function: Grid of conditional probabilities}
\usage{
grid.lik(xigrid, mutns, probamut)
}
\arguments{
\item{xigrid}{Grid of the values of xi, corresponding to its domain of definition.}

\item{mutns}{Matrix of the mutations observed, with all mutations in rows and the cases (tumor pairs) in columns. 
The data are coded as 0=mutation not observed, 1=shared mutation (observed in both tumors), 2=private mutation 
(observed in one tumor only).}

\item{probamut}{Vector of the probabilities of occurence for each mutation.}
}
\value{
Return the matrix of the likelihood values for each tumor pair (rows) and each value of xi (columns). This matrix is called 
by the auxiliary function grid.lik, returned as a parameter by the function clonal.est, and used as a parameter by the function clonal.proba.
}
\description{
This auxiliary function generates the grid of likelihood values for each tumor pair (rows) and each value of xi (columns): P(observed mutations | xi)
}

