% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMGT_Parse.R
\docType{data}
\name{imgt.header}
\alias{imgt.header}
\title{Annotate clonality}
\format{
An object of class \code{character} of length 33.
}
\usage{
imgt.header
}
\arguments{
\item{data}{A data frame object or the full path to a xlsx with repertoire data.}
}
\description{
Returns the input table with the clonality annotations.
}
\keyword{datasets}
