% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.lik.r
\name{model.lik}
\alias{model.lik}
\title{Auxiliary likelihood Function}
\usage{
model.lik(para, likmat, out0, xigrid)
}
\arguments{
\item{para}{Value of the model parameters, in the form c(mu, sigma, pi).}

\item{likmat}{Grid of conditional probabilities for each tumor pair (rows) and each value of xi (columns). 
This matrix is generated by the function grid.lik.}

\item{out0}{a small value that is used when the likelihood goes to infinite values, posing problem for 
the maximization. The corresponding combination of the parameters will thus be excluded from the search.}

\item{xigrid}{Grid of the values of xi, corresponding to its domain of definition.}
}
\value{
Return the likelihood value of the model for the given parameters. This likelihood function is 
the one that is maximized in the clonal.est function.
}
\description{
This function computes the likelihood of the model.
}

