% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mutation.proba.r
\name{print.mutation.proba}
\alias{print.mutation.proba}
\title{Print for the mutation.proba function}
\usage{
## S3 method for class 'mutation.proba'
print(x, ...)
}
\arguments{
\item{x}{a mutation.proba object}

\item{...}{Other unused arguments.}
}
\value{
Print results for the individual probabilities of clonality.
}
\description{
Print a summary of results for the probabilities of clonality estimated by the mutation.proba function.
}
\seealso{
mutation.proba
}

