% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mutation.rem.r
\name{print.mutation.rem}
\alias{print.mutation.rem}
\title{Print for the mutation.rem function}
\usage{
## S3 method for class 'mutation.rem'

\method{print}{mutation.rem}(x, ...)
}
\arguments{
\item{x}{a mutation.rem object}

\item{...}{Other unused arguments.}
}
\value{
Print results for the model estimates.
}
\description{
Print a summary of results for the random-effect model estimation estimated by the clonal.est function.
}
\seealso{
mutation.rem
}

