% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_genenames.R
\name{simplify_genenames}
\alias{simplify_genenames}
\title{Simplify Gene Names Based on Cell Type and Create a Summary Report}
\usage{
simplify_genenames(data, cell)
}
\arguments{
\item{data}{A data frame containing immune repertoire data with V and J gene columns.}

\item{cell}{\code{Character}. The cell type: "T" for T cells or "B" for B cells. The function adjusts the gene name extraction pattern accordingly.}
}
\value{
A list with:
\itemize{
\item \code{clonal.df}: A data frame with simplified gene names and added metadata columns.
\item \code{report}: A summary data frame with counts of functional types and multiple annotations.
}
}
\description{
This helper function simplifies V and J gene names in the input data frame according to the specified cell type (T cell or B cell).
It also generates a report summarizing the number of functional ("F"), open reading frame ("ORF"), and pseudo ("P") genes,
as well as sequences with multiple annotations.
}
\examples{
data <- data.frame(v_genes = c("TRAV12-2*01 F", "TRBV7-9*01 F"),
                   j_genes = c("TRAJ33*01 F", "TRBJ2-1*01 F"))
result <- simplify_genenames(data, cell = "T")
simplified <- result$clonal.df
report <- result$report
}
