% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{tenx}
\alias{tenx}
\title{Annotate 10x clones}
\usage{
tenx(
  data = NULL,
  method = "unique_paired",
  only_productive = T,
  clonality_input = NULL,
  cell = "T",
  save.files = F
)
}
\arguments{
\item{data}{Character. Data frame object or the full path to a filtered_contig_annotations.csv file.}

\item{method}{Character. One of: unique_paired, unique_all, sticky_ends. Default: unique_paired.}

\item{only_productive}{Logical. Filter non productive chains.}

\item{clonality_input}{Named vector. Input parameters for the clonality function.}

\item{cell}{Character. Possible values: \code{B} Bcells, \code{T} Tcells, \code{Tgd} Gamma Delta T cells.}
}
\description{
Returns the input with clonality annotation column with clonal definitions.
}
\examples{
tenx(data = "filtered_contig_annotations", method = "sticky_ends", only_productive = T, clonality_input = c("mm" = 0.25), cell = "T",  save.files = F)
}
