% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_input.R
\name{test_input}
\alias{test_input}
\title{Test input}
\usage{
test_input(test_data = NULL)
}
\arguments{
\item{test_data}{A data frame object or the full path to a xlsx with repertoire data.}
}
\description{
Evaluate if input data have the required artributes.
}
