% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xidens.r
\name{xidens}
\alias{xidens}
\title{Auxiliary function computing the density of xi}
\usage{
xidens(pmu, psig, xigrid)
}
\arguments{
\item{pmu}{Mean parameter of the distribution.}

\item{psig}{Variance parameter of the distribution.}

\item{xigrid}{Grid of the values of xi, corresponding to its domain of definition.}
}
\value{
Returns the density value for the given values of xi.
}
\description{
Density function for the random variable xi, using a lognormal density for phi=-log(1-xi)
}

