% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seurat_metadata.R
\name{add_seurat_metadata}
\alias{add_seurat_metadata}
\title{Define 10x clonotypes}
\usage{
add_seurat_metadata(
  seu = NULL,
  clonal.list = ls(pattern = "^Clonal", envir = .GlobalEnv),
  cdr3_map = "cdr3_col2",
  purity = 0.8
)
}
\arguments{
\item{seu}{Seurat. A seurat object to embed with clonality data.}

\item{clonal.list}{Character. The clonalities to be inserted into the seurat metadata. Default:  \code{ls(pattern = "^Clonal")}.}

\item{cdr3_map}{Character. The columun name to be used to map single chains to paired chains. CDR3 for nt ; cdr3_col2 for aa. Default: "cdr3_col2".}

\item{purity}{Numeric. Minimum fraction of the dominant mapped paired to be assigned to single.}
}
\description{
Takes a list of paired chains and defines clonotypes.
}
