\name{Clonality-package}
\alias{Clonality-package}
\alias{Clonality}
\docType{package}
\title{
Clonality testing
}
\description{
Statistical tests for clonality versus independence of tumors from the same patient based on their LOH or genomewide copy number profiles.
}
\details{
\tabular{ll}{
Package: \tab Clonality\cr
Type: \tab Package\cr
Version: \tab 0.99.3 \cr
Date: \tab 2014-9-07\cr
License: \tab  GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
          
Irina Ostrovnaya \email{ostrovni@mskcc.org}
}
