% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{clonality}
\alias{clonality}
\title{Define clonality}
\usage{
clonality(
  File = "example.xlsx",
  NewF = "output",
  ID_Column = "Sequence ID",
  Vgene_Column = "V-GENE and allele",
  Jgene_Column = "J-GENE and allele",
  CDR3_Column = "JUNCTION",
  cell = "T",
  rm.junc.na = TRUE,
  output.orig = FALSE,
  Mismatch = 0
)
}
\arguments{
\item{File}{Data frame or full path to a xlsx input containing TCR/BCR repertoire data.}

\item{NewF}{The file output name.}

\item{ID_Column}{The column containing an individual sequence ID.}

\item{Vgene_Column}{The column name containing V gene names.}

\item{Jgene_Column}{The column name containing J gene names.}

\item{CDR3_Column}{The column name containing CDR3 sequences - nt or aa.}

\item{cell}{Choose between T for T cell or B for Bcell.}

\item{rm.junc.na}{Remove rows with NA junctions. If \code{FALSE}, NA rows are considered unique sequences.}

\item{output.orig}{If \code{TRUE}, outputs the original input, if \code{FALSE} outputs a minimum table.}

\item{Mismatch}{Percent of mismatches allowed in CDR3 before subsetting a group.}
}
\description{
Returns the input with an additional clonality column with clonal definitions.
}
\examples{
clonality(File = tra)
clonality(File = trb)
clonality(File = "example.xlsx")
}
