% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonality.R
\name{Clonality}
\alias{Clonality}
\title{Define clonality}
\usage{
Clonality(
  File = "input.xlsx",
  NewF = "output.xlsx",
  Vgene_Column = "V-GENE and allele",
  Jgene_Column = "J-GENE and allele",
  CDR3_Column = "JUNCTION",
  cell = "T",
  rm.junc.na = TRUE,
  Mismatch = 0
)
}
\arguments{
\item{File}{Path to a xlsx input containing the clonality data to be defined.}

\item{NewF}{The output name.}

\item{Vgene_Column}{The column name containing V gene names.}

\item{Jgene_Column}{The column name containing J gene names.}

\item{CDR3_Column}{The column name containing CDR3 sequences - nt or aa.}

\item{cell}{Choose between T for T cell or B for Bcell.}

\item{rm.junc.na}{Remove rows with NA junctions. If \code{FALSE}, NA rows are considered unique.}

\item{Mismatch}{Percent of mismatches allowed in CDR3 before subsetting a group.}
}
\description{
Returns the input with an additional clonality column with clonal definitions.
}
