% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{clonality}
\alias{clonality}
\title{Annotate clonality}
\usage{
clonality(
  data = "example.xlsx",
  output = "output",
  ident_col = "Sequence_ID",
  vgene_col = "V_GENE_and_allele",
  jgene_col = "J_GENE_and_allele",
  cdr3_col = "JUNCTION",
  cell = "T",
  output_original = FALSE,
  mismatch = 0,
  suffix = NULL,
  search_genename = TRUE,
  stringdist_method = "osa"
)
}
\arguments{
\item{data}{A data frame object or the full path to a xlsx with repertoire data.}

\item{output}{\code{Character}. The output name. Default: \code{output.}}

\item{ident_col}{\code{Character}. The column with unique sequence IDs. Default: \code{Sequence_ID}.}

\item{vgene_col}{\code{Character}. The column with V gene names. Default: \code{V_GENE_and_allele}.}

\item{jgene_col}{\code{Character}. The column with J gene names. Default: \code{J_GENE_and_allele}.}

\item{cdr3_col}{\code{Character}. The column with CDR3 sequences - nt/aa. Default: \code{JUNCTION}.}

\item{cell}{\code{Character}. Choose T for T cell or B for B cell. Default: \code{T}.}

\item{output_original}{\code{Logical}. if\code{TRUE}, output and input have same format. Default: \code{FALSE}.}

\item{mismatch}{\code{Numeric.} Percent of mismatches allowed in subgroups. \code{0 == most_stringent}, \code{1 == less_stringent}. Default: \code{0}}

\item{suffix}{\code{Character.} String to be appended to the clonality ID.}

\item{search_genename}{\code{Logical.} Accepts IMGT nomenclature and simplify gene IDs. Default: \code{TRUE}.}

\item{stringdist_method}{\code{Character.} Method for distance calculation. See \code{?stringdist-metrics} for more methods.}
}
\description{
The \pkg{clonality} package returns the input table with the clonality annotations.
}
\examples{
clonality(data = tra)
clonality(data = trb)
clonality(data = 'example.xlsx')
}
