% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_immunarch.R
\name{gen_immunarch}
\alias{gen_immunarch}
\title{Generates an immunarch compatible input}
\usage{
gen_immunarch(seu = seu, proportion = "detected", sample_column = NULL)
}
\arguments{
\item{proportion}{Character. Possible values: \code{all} against all cells, \code{detected} remove NAs prior to proportion.}

\item{sample_column}{Character. Column to be used as sample to cut the data.}

\item{data}{A data frame object or the full path to a xlsx with repertoire data.}
}
\description{
Returns a dataset compatible with the immunarch package.
}
\examples{
gen_immunarch(seu, proportion = "detected",  sample_column = "Conditions")
}
