\name{lcis}
\alias{lcis}

\title{
  Breast cancer data
}
\description{
  For each sample the log-ratios at each consecutive K number of probes are averaged.
}
\usage{
  data(lcis
}

\details{
  This is exome sequencing data from study of Lobular Carcinoma in Situ (LCIS) and  
  Invaisve lobular carcinomas  (ILC) or Invasive Ductal Carcinomas (IDC) in the same patients.
  First column called probi contains marginal probabilities that   are obtained from breast cancer TCGA data and are not directly applicable to other cancers.
  Each subsequent column contains a pair of tumors where value 0  denotes that mutation is not 
observed, 1 if shared mutation is observed in both tumors, and 2 if it is a private mutation observed in only one tumor.
}
\references{
  Begg CB, Ostrovnaya I, Carniello JV, Sakr RA, Giri D, Towers R, Schizas M, De Brot M, Andrade VP, Mauguen A, Seshan VE, King TA. "Clonal relationships between lobular carcinoma in situ and other breast malignancies.", Breast Cancer Res. 2016 Jun 23;18(1):66. doi: 10.1186/s13058-016-0727-z.

}

