% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.rem.r
\name{mutation.rem}
\alias{mutation.rem}
\title{Estimation of the random-effect model for clonality based on mutations.}
\usage{
mutation.rem(
  mutmat,
  proba = FALSE,
  print.proba = FALSE,
  xigrid = seq(5e-04, 0.9995, by = 0.001),
  init.para = c(0, 1, 0.5),
  conv.crit = 1e-05,
  niter = 300
)
}
\description{
Estimation of the random-effect model for clonality based on mutations.
}
