% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{tenx}
\alias{tenx}
\title{Define 10x genomics clonotypes}
\usage{
tenx(
  data = NULL,
  method = "unique_paired",
  only_productive = T,
  clonality_input = NULL,
  cell = "T",
  pairing = "_",
  save.files = F
)
}
\arguments{
\item{data}{Data frame object or the full path to a filtered_contig_annotations.csv input.}

\item{method}{One of: unique_paired, unique_all, sticky_ends. Default: unique_paired. Character.}

\item{only_productive}{Filter non productive chains. Logical.}

\item{clonality_input}{Input parameters for the clonality function. List.}
}
\description{
Returns the input with an additional clonality column with clonal definitions.
}
\examples{
clonality(data = filtered_contig_annotations)
}
