% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{tenx}
\alias{tenx}
\title{Define 10x clonotypes}
\usage{
tenx(
  data = NULL,
  method = "unique_paired",
  only_productive = T,
  clonality_input = NULL,
  cell = "T",
  save.files = F
)
}
\arguments{
\item{data}{Data frame object or the full path to a filtered_contig_annotations.csv input.}

\item{method}{One of: unique_paired, unique_all, sticky_ends. Default: unique_paired. Character.}

\item{only_productive}{Filter non productive chains. Logical.}

\item{clonality_input}{Input parameters for the clonality function. List.}

\item{cell}{Character. Possible values: \code{B} Bcells, \code{T} Tcells, \code{Tgd} Tcells GamaDelta.}
}
\description{
Returns the input with an additional clonality column with clonal definitions.
}
\examples{
tenx(data = "filtered_contig_annotations", method = "sticky_ends", only_productive = T, clonality_input = c("mm" = 0.25), cell = "T",  save.files = F)
}
