% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{tenx}
\alias{tenx}
\title{Define tenx}
\usage{
tenx(File = "example.xlsx", method = "unique_paired")
}
\arguments{
\item{File}{Data frame object or the full path to a xlsx input containing TCR/BCR repertoire data.}

\item{method}{One of: unique_paired, unique_all, sticky_ends. Default: unique_paired.}
}
\description{
Returns the input with an additional clonality column with clonal definitions.
}
\examples{
clonality(File = tra)
}
